requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

albucore==0.0.24
albumentations==2.0.7
annotated-types==0.7.0
colorama==0.4.6
eval-type-backport==0.2.2
filelock==3.13.1
flowio==1.4.0
fsspec==2024.6.1
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
pillow==11.0.0
pydantic==2.11.4
pydantic-core==2.33.2
PyYAML==6.0.2
scipy==1.13.1
simsimd==6.2.1
stringzilla==3.12.5
sympy==1.13.3
torch==2.7.0+cu126
torchvision==0.22.0+cu126
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
scikit-image==0.24.0
